--- client/src/components/Login.jsx
@@ -1,27 +1,0 @@
-import React from 'react'
-import "../styles/login.css"
-
-export const Login = () => {
-  return (
-    <div className="login">
-      <span className="loginTitle">Login</span>
-      <form className="loginForm">
-        <label>Username</label>
-        <input
-          type="text"
-          className="loginInput"
-          placeholder="Enter your username..."
-        />
-        <label>Password</label>
-        <input
-          type="password"
-          className="loginInput"
-          placeholder="Enter your password..."
-        />
-        <button className="loginButton" type="submit">
-          Login
-        </button>
-      </form>
-    </div>
-  );
-}
